refactor: 주문 체결 DTO 클래스 위치 수정
