+++ src/main/java/com/devspacehub/ast/domain/my/service/MyServiceImpl.java
@@ -11,14 +11,14 @@ package com.devspacehub.ast.domain.my.service;
 import com.devspacehub.ast.common.config.OpenApiProperties;
 import com.devspacehub.ast.common.constant.OpenApiType;
 import com.devspacehub.ast.domain.my.dto.MyServiceRequestDto;
-import com.devspacehub.ast.domain.my.orderConclusion.dto.OrderConclusionFindExternalReqDto;
-import com.devspacehub.ast.domain.my.orderConclusion.dto.OrderConclusionFindExternalResDto;
+import com.devspacehub.ast.domain.my.dto.orderConclusion.OrderConclusionFindExternalReqDto;
+import com.devspacehub.ast.domain.my.dto.orderConclusion.OrderConclusionFindExternalResDto;
 import com.devspacehub.ast.domain.my.reservationOrderInfo.ReservationOrderInfoRepository;
 import com.devspacehub.ast.domain.my.stockBalance.dto.request.StockBalanceApiReqDto;
 import com.devspacehub.ast.domain.my.stockBalance.dto.response.BuyPossibleCashApiResDto;
 import com.devspacehub.ast.domain.my.stockBalance.dto.response.StockBalanceApiResDto;
 import com.devspacehub.ast.domain.my.stockBalance.dto.request.BuyPossibleCashApiReqDto;
-import com.devspacehub.ast.domain.orderTrading.dto.OrderConclusionDto;
+import com.devspacehub.ast.domain.my.dto.orderConclusion.OrderConclusionDto;
 import com.devspacehub.ast.exception.error.OpenApiFailedResponseException;
 import com.devspacehub.ast.util.OpenApiRequest;
 import lombok.extern.slf4j.Slf4j;
