+++ src/test/java/com/devspacehub/ast/domain/my/service/MyServiceImplIntegrationTest.java
@@ -13,8 +13,7 @@ import com.devspacehub.ast.common.constant.YesNoStatus;
 import com.devspacehub.ast.domain.my.reservationOrderInfo.QReservationOrderInfo;
 import com.devspacehub.ast.domain.my.reservationOrderInfo.ReservationOrderInfo;
 import com.devspacehub.ast.domain.my.reservationOrderInfo.ReservationOrderInfoRepository;
-import com.devspacehub.ast.domain.my.stockBalance.dto.response.StockBalanceApiResDto;
-import com.devspacehub.ast.domain.orderTrading.dto.OrderConclusionDto;
+import com.devspacehub.ast.domain.my.dto.orderConclusion.OrderConclusionDto;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
